refactor: clean up unused relationships and imports in database models; update documentation structure
