+++ api/pyproject.toml
@@ -17,7 +17,6 @@ dependencies = [
     'asyncpg>=0.31.0',
     'botocore>=1.40.61',
     'cryptography>=50.0.0',
-    'email-validator>=2.3.0',
     'exoscale>=0.16.3',
     'fastapi>=0.141.1',
     'httpx2>=2.9.1',
