+++ web/src/platform/admin/pagination.ts
@@ -4,10 +4,7 @@ import { paginateData, useTablePagination } from '@astryxdesign/core/Table';
 const PAGE_SIZE = 25;
 
 /** Provides consistent client-side pagination for Platform admin tables. */
-export function useAdminPagination<T extends Record<string, unknown>>(
-    items: T[],
-    controls: 'compact' | 'default' = 'compact'
-) {
+export function useAdminPagination<T extends Record<string, unknown>>(items: T[]) {
     const [page, setPage] = useState(1);
     const currentPage = Math.min(page, Math.max(1, Math.ceil(items.length / PAGE_SIZE)));
 
@@ -16,12 +13,8 @@ export function useAdminPagination<T extends Record<string, unknown>>(
         onPageChange: setPage,
         totalItems: items.length,
         pageSize: PAGE_SIZE,
-        ...(controls === 'compact'
-            ? {
-                  label: 'Previous / Next',
-                  size: 'sm' as const,
-              }
-            : {}),
+        label: 'Previous / Next',
+        size: 'sm',
     });
 
     return {
