+++ web/src/platform/legal/catalog.tsx
@@ -1,14 +1,13 @@
 import { createElement } from 'react';
 import type { ArticlePage } from '@/lib/articles';
-import { legalPages } from '@/platform/legal/pages';
+import { legalPages } from '@/platform/pages';
 
-export const LEGAL_PAGES: ArticlePage[] = legalPages.map(({ Component, Icon, metadata }) => ({
+export const LEGAL_PAGES: ArticlePage[] = legalPages.map(({ Component, metadata }) => ({
     ...metadata,
     breadcrumbs: [
         { title: 'Home', path: '/' },
         { title: metadata.title, path: metadata.path },
     ],
-    icon: createElement(Icon, { 'aria-hidden': true, size: 16 }),
     content: createElement(Component),
     metadata,
 }));
@@ -16,10 +15,6 @@ export const LEGAL_PAGES: ArticlePage[] = legalPages.map(({ Component, Icon, met
 export const LEGAL_GROUPS = [
     {
         title: 'Legal',
-        items: LEGAL_PAGES.map((page) => ({
-            title: page.title,
-            path: page.path,
-            icon: page.icon,
-        })),
+        items: LEGAL_PAGES.map(({ title, path }) => ({ title, path })),
     },
 ];
